+++ tests/back/Pim/Enrichment/Integration/Classification/ClassifyProductModelIntegration.php
@@ -2,8 +2,11 @@
 
 namespace AkeneoTest\Pim\Enrichment\Integration\Classification;
 
+use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
+use Akeneo\Pim\Enrichment\Product\API\Command\UpsertProductCommand;
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\SetCategories;
+use Akeneo\Pim\Enrichment\Product\API\Command\UserIntent\UserIntent;
 use Akeneo\Test\Integration\TestCase;
-use AkeneoTest\Pim\Enrichment\Integration\Fixture\EntityBuilder;
 
 class ClassifyProductModelIntegration extends TestCase
 {
@@ -71,10 +74,7 @@ class ClassifyProductModelIntegration extends TestCase
 
     public function testClassifyChildrenProductDoesNotImpactItsParent()
     {
-        $childrenProduct = $this->get('pim_catalog.repository.product')->findOneByIdentifier('tee-red-s');
-        $this->get('pim_catalog.updater.product')->update($childrenProduct, ['categories' => ['master_women_shirts']]);
-        $this->get('pim_catalog.validator.product')->validate($childrenProduct);
-        $this->get('pim_catalog.saver.product')->save($childrenProduct);
+        $this->createProduct('tee-red-s', [new SetCategories(['master_women_shirts'])]);
 
         $parentProductModel = $this->get('pim_catalog.repository.product_model')->findOneByIdentifier('model-tee-red');
         $this->assertCount(1, $parentProductModel->getCategories());
@@ -84,10 +84,8 @@ class ClassifyProductModelIntegration extends TestCase
 
     public function testClassifyWithSameCategoryThanParentHasNoImpactOnChildrenProduct()
     {
+        $this->createProduct('tee-red-s', [new SetCategories(['supplier_zaro'])]);
         $childrenProduct = $this->get('pim_catalog.repository.product')->findOneByIdentifier('tee-red-s');
-        $this->get('pim_catalog.updater.product')->update($childrenProduct, ['categories' => ['supplier_zaro']]);
-        $this->get('pim_catalog.validator.product')->validate($childrenProduct);
-        $this->get('pim_catalog.saver.product')->save($childrenProduct);
 
         $this->assertCount(1, $childrenProduct->getCategories());
         $category = $childrenProduct->getCategories()->first();
@@ -96,10 +94,7 @@ class ClassifyProductModelIntegration extends TestCase
 
     public function testUnclassifyChildrenProductDoesNotImpactItsParents()
     {
-        $childrenProduct = $this->get('pim_catalog.repository.product')->findOneByIdentifier('tee-red-s');
-        $this->get('pim_catalog.updater.product')->update($childrenProduct, ['categories' => []]);
-        $this->get('pim_catalog.validator.product')->validate($childrenProduct);
-        $this->get('pim_catalog.saver.product')->save($childrenProduct);
+        $this->createProduct('tee-red-s', [new SetCategories([])]);
 
         $parentProductModel = $this->get('pim_catalog.repository.product_model')->findOneByIdentifier('model-tee');
         $this->assertCount(1, $parentProductModel->getCategories());
@@ -151,5 +146,34 @@ class ClassifyProductModelIntegration extends TestCase
     protected function getConfiguration()
     {
         return $this->catalog->useFunctionalCatalog('catalog_modeling');
+    }
+
+    /**
+     * @param UserIntent[] $userIntents
+     */
+    protected function createProduct(string $identifier, array $userIntents): ProductInterface
+    {
+        $command = UpsertProductCommand::createFromCollection(
+            userId: $this->getUserId('admin'),
+            productIdentifier: $identifier,
+            userIntents: $userIntents
+        );
+        $this->get('pim_enrich.product.message_bus')->dispatch($command);
+
+        return $this->get('pim_catalog.repository.product')->findOneByIdentifier($identifier);
+    }
+
+    protected function getUserId(string $username): int
+    {
+        $query = <<<SQL
+            SELECT id FROM oro_user WHERE username = :username
+        SQL;
+        $stmt = $this->get('database_connection')->executeQuery($query, ['username' => $username]);
+        $id = $stmt->fetchOne();
+        if (null === $id) {
+            throw new \InvalidArgumentException(\sprintf('No user exists with username "%s"', $username));
+        }
+
+        return \intval($id);
     }
 }